Use s32 and u32 instead of long and unsigned long

+++ include/Camera/SunMgr.hpp
@@ -13,7 +13,7 @@ public:
 	TSunMgr(const char* name = "<TSunMgr>");
 
 	virtual void load(JSUMemoryInputStream&);
-	virtual void perform(unsigned long, JDrama::TGraphics*);
+	virtual void perform(u32, JDrama::TGraphics*);
 	virtual void drawSyncCallback(unsigned short);
 
 	int getAddColor() const;
